+++ branch-alpha/.ci.sh
@@ -542,7 +542,6 @@ shSqlmathUpdate() {(set -e
         for FILE in \
             .ci.sh \
             asset_sqlmath_external_rollup.js \
-            indent.exe \
             index.html \
             setup.py \
             sqlmath.mjs \
@@ -551,7 +550,8 @@ shSqlmathUpdate() {(set -e
             sqlmath_base.h \
             sqlmath_browser.mjs \
             sqlmath_external_sqlite.c \
-            sqlmath_wrapper_wasm.js
+            sqlmath_wrapper_wasm.js \
+            zlib.v1.3.1.vcpkg.x64-windows-static.lib
         do
             ln -f "$HOME/Documents/sqlmath/$FILE" "$FILE"
         done