+++ branch-alpha/sqlmath_base.c
@@ -113,7 +113,7 @@ file sqlmath_h - start
 
 
 #if !defined(SQLITE_MAX_FUNCTION_ARG)
-#   define SQLITE_MAX_FUNCTION_ARG 127
+#   define SQLITE_MAX_FUNCTION_ARG 1000
 #endif                          // SQLITE_MAX_FUNCTION_ARG
 
 
@@ -1691,71 +1691,77 @@ SQLMATH_FUNC static void sql1_gzip_compress_func(
     int argc,
     sqlite3_value ** argv
 ) {
-// This function will gzip-compress <argv[0]>
-// using miniz's deflate-algorithm.
+// This function will gzip-compress <argv[0]> using zlib's deflate-algorithm.
     UNUSED_PARAMETER(argc);
-    static const char header[10] = {    //
-        0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03,     //
-    };
-    const mz_uint8 *buf_src = sqlite3_value_blob(argv[0]);
-    if (buf_src == NULL) {
+    // declare var
+    z_stream strm = { 0 };
+    int errcode = Z_OK;
+    uint8_t *gzip_buf = NULL;
+    // init argv
+    const uint8_t *src_buf = sqlite3_value_blob(argv[0]);
+    if (src_buf == NULL) {
         sqlite3_result_error(context,   //
-            "gzip_compress: Input cannot be NULL", -1);
+            "gzip_compress - Input cannot be NULL", -1);
         return;
     }
-    const uint32_t len_src = sqlite3_value_bytes(argv[0]);
-    if (SIZEOF_BLOB_MAX < len_src) {
+    const int src_len = sqlite3_value_bytes(argv[0]);
+    if (SIZEOF_BLOB_MAX < src_len) {
         goto catch_error;
     }
-    // Part 1: Compute CRC32 and store original size
-    // mz_ulong mz_crc32(
-    //     mz_ulong crc,
-    //     const mz_uint8 *ptr,
-    //     size_t buf_len
+    // uLong compressBound(
+    //     uLong sourceLen
     // );
-    const uint32_t crc = mz_crc32(MZ_CRC32_INIT, buf_src, (size_t) len_src);
-    // Handle zero-length input separately to produce a valid 18-byte gzip file
-    if (len_src == 0) {
-        char buf_gzip[10 + 0 + 8] = {   //
-            0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, //
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        };
-        memcpy(buf_gzip + 10, &crc, 4);
-        memcpy(buf_gzip + 14, &len_src, 4);
-        sqlite3_result_blob(context, buf_gzip, 10 + 0 + 8, SQLITE_TRANSIENT);
-        return;
+    const int gzip_len = (int) compressBound(src_len) + 18;
+    gzip_buf = sqlite3_malloc(gzip_len);
+    if (!gzip_buf) {
+        goto catch_error;
     }
-    // Part 2: Perform Deflate compression with miniz.
-    // tdefl_compress_mem_to_heap produces a raw Deflate stream, which is
-    // precisely what is needed for the gzip format.
-    size_t len_compress = 0;
-    // void *tdefl_compress_mem_to_heap(
-    //     const void *pSrc_buf,
-    //     size_t src_buf_len,
-    //     size_t * pOut_len,
-    //     int flags
+    // int deflateInit2(
+    //     z_streamp strm,
+    //     int level,
+    //     int method,
+    //     int windowBits,
+    //     int memLevel,
+    //     int strategy
     // );
-    char *buf_gzip =
-        tdefl_compress_mem_to_heap(buf_src, (size_t) len_src, &len_compress,
-        0);
-    if (buf_gzip == NULL) {
+    errcode = deflateInit2(     //
+        &strm,                  //
+        Z_DEFAULT_COMPRESSION,  //
+        Z_DEFLATED,             //
+        15 + 16,                //
+        8,                      //
+        Z_DEFAULT_STRATEGY);
+    if (errcode != Z_OK) {
         goto catch_error;
     }
-    // Part 3: Construct the full gzip buffer.
-    // Gzip Header (10 bytes) + Compressed Data + Gzip Footer (8 bytes)
-    buf_gzip = realloc(buf_gzip, 10 + len_compress + 8);
-    if (buf_gzip == NULL) {
-        goto catch_error;
+    strm.avail_in = (uLong) src_len;
+    strm.next_in = (uint8_t *) src_buf;
+    strm.avail_out = (uLong) gzip_len;
+    strm.next_out = gzip_buf;
+    // int deflate(
+    //     z_streamp strm,
+    //     int flush
+    // );
+    errcode = deflate(&strm, Z_FINISH);
+    if (errcode != Z_STREAM_END) {
+        sqlite3_result_error2(context,  //
+            "gzip_compress - deflate() failed - %d\n", errcode);
+        goto cleanup;
     }
-    memmove(buf_gzip + 10, buf_gzip, len_compress);
-    // Copy the header, compressed data, and footer to the final buffer
-    memcpy(buf_gzip + 0, header, 10);
-    memcpy(buf_gzip + 10 + len_compress, &crc, 4);
-    memcpy(buf_gzip + 10 + len_compress + 4, &len_src, 4);
-    sqlite3_result_blob(context, buf_gzip, 10 + (int) len_compress + 8, free);
+    sqlite3_result_blob(context, gzip_buf, gzip_len - (int) strm.avail_out,
+        sqlite3_free);
+    deflateEnd(&strm);
     return;
   catch_error:
     sqlite3_result_error_nomem(context);
+  cleanup:
+    // int deflateEnd(
+    //     z_streamp strm
+    // );
+    deflateEnd(&strm);
+    if (gzip_buf) {
+        sqlite3_free(gzip_buf);
+    }
 }
 
 SQLMATH_FUNC static void sql1_gzip_uncompress_func(
@@ -1763,68 +1769,76 @@ SQLMATH_FUNC static void sql1_gzip_uncompress_func(
     int argc,
     sqlite3_value ** argv
 ) {
-// This function will gzip-uncompress <argv[0]>
-// using miniz's inflate-algorithm.
+// This function will gzip-uncompress <argv[0]> using zlib's inflate-algorithm.
     UNUSED_PARAMETER(argc);
-    static const char header[4] = { 0x1f, 0x8b, 0x08, 0x00 };
-    const char *buf_gzip = sqlite3_value_blob(argv[0]);
-    if (buf_gzip == NULL) {
+    // declare var
+    z_stream strm = { 0 };
+    uint8_t *src_buf = NULL;
+    int src_len = 4096;
+    int errcode = Z_OK;
+    // init argv
+    const uint8_t *gzip_buf = sqlite3_value_blob(argv[0]);
+    if (gzip_buf == NULL) {
         sqlite3_result_error(context,   //
-            "gzip_uncompress: Input cannot be NULL", -1);
+            "gzip_uncompress - Input cannot be NULL", -1);
         return;
     }
-    const size_t len_gzip = sqlite3_value_bytes(argv[0]);
-    if (SIZEOF_BLOB_MAX < len_gzip) {
+    const int gzip_len = sqlite3_value_bytes(argv[0]);
+    if (SIZEOF_BLOB_MAX < gzip_len) {
         sqlite3_result_error_nomem(context);
         return;
     }
-    // Check for minimum gzip file size (10 byte header + 8 byte footer)
-    if (len_gzip < 18) {
-        sqlite3_result_error(context,   //
-            "gzip_uncompress: Invalid gzip format, buffer too small", -1);
-        return;
-    }
-    // Validate the gzip header
-    if (memcmp(buf_gzip, header, 4) != 0) {
-        sqlite3_result_error(context,   //
-            "gzip_uncompress: Invalid gzip header or unsupported flags", -1);
-        return;
-    }
-    // Extract compressed data, CRC32, and original size from the buffer
-    // Decompress the data
-    size_t len_src = 0;
-    // void *tinfl_decompress_mem_to_heap(
-    //     const void *pSrc_buf,
-    //     size_t src_buf_len,
-    //     size_t * pOut_len,
-    //     int flags
+    // int inflateInit2(
+    //     z_streamp strm,
+    //     int windowBits
     // );
-    mz_uint8 *buf_src =
-        tinfl_decompress_mem_to_heap(buf_gzip + 10, len_gzip - 18, &len_src,
-        0);
-    if (buf_src == NULL) {
-        sqlite3_result_error(context,   //
-            "gzip_uncompress: Decompression failed", -1);
-        return;
+    errcode = inflateInit2(&strm, 15 + 16);
+    if (errcode != Z_OK) {
+        goto catch_error;
     }
-    // Check for CRC and LEN match
-    uint32_t crc_header;
-    uint32_t len_header;
-    memcpy(&crc_header, buf_gzip + len_gzip - 8, 4);
-    memcpy(&len_header, buf_gzip + len_gzip - 4, 4);
-    // mz_ulong mz_crc32(
-    //     mz_ulong crc,
-    //     const mz_uint8 *ptr,
-    //     size_t buf_len
+    strm.avail_in = (uLong) gzip_len;
+    strm.next_in = (uint8_t *) gzip_buf;
+    // Allocate the initial output buffer
+    src_buf = sqlite3_malloc(src_len);
+    if (!src_buf) {
+        goto catch_error;
+    }
+    do {
+        // Set up the output buffer for the next decompression step.
+        // `total_out` is the total bytes decompressed so far.
+        strm.avail_out = (uLong) src_len - strm.total_out;
+        strm.next_out = src_buf + strm.total_out;
+        errcode = inflate(&strm, Z_NO_FLUSH);
+        if (errcode < 0 && errcode != Z_STREAM_END) {
+            sqlite3_result_error2(context,      //
+                "gzip_uncompress - inflate() failed - %d\n", errcode);
+            goto cleanup;
+        }
+        // If the output buffer is full, reallocate to double its capacity.
+        if (strm.avail_out == 0 && errcode != Z_STREAM_END) {
+            src_len = src_len * 2;
+            if (SIZEOF_BLOB_MAX < src_len) {    // Check for overflow
+                goto catch_error;
+            }
+            src_buf = sqlite3_realloc(src_buf, src_len);
+            if (!src_buf) {
+                goto catch_error;
+            }
+        }
+    } while (errcode != Z_STREAM_END);
+    sqlite3_result_blob(context, src_buf, (int) strm.total_out, sqlite3_free);
+    inflateEnd(&strm);
+    return;
+  catch_error:
+    sqlite3_result_error_nomem(context);
+  cleanup:
+    // int inflateEnd(
+    //     z_streamp strm
     // );
-    uint32_t crc_actual = mz_crc32(MZ_CRC32_INIT, buf_src, len_src);
-    if (!(crc_header == crc_actual && len_header == len_src)) {
-        free(buf_src);
-        sqlite3_result_error(context,   //
-            "gzip_uncompress: CRC or uncompressed size mismatch", -1);
-        return;
+    inflateEnd(&strm);
+    if (src_buf) {
+        sqlite3_free(src_buf);
     }
-    sqlite3_result_blob(context, buf_src, (int) len_src, free);
 }
 
 // SQLMATH_FUNC sql1_gzip_xxx_func - end
@@ -2055,6 +2069,7 @@ SQLMATH_FUNC static void sql1_lgbm_dlopen_func(
     int errcode = 0;
     errcode = dbDlopen(context, filename, (void **) &lgbm_library);
     if (errcode) {
+        sqlite3_result_error_code(context, errcode);
         return;
     }
     LGBM_dlsym();
@@ -2240,13 +2255,13 @@ SQLMATH_FUNC static void sql1_lgbm_predictforfile_func(
     sqlite3_value ** argv
 ) {
 // This function will make prediction for file from <model_str>.
+    UNUSED_PARAMETER(argc);
     const char *model_str = (char *) sqlite3_value_text(argv[0]);
     if (model_str == NULL) {
         sqlite3_result_error(context,
             "lgbm_predictforfile - model_str cannot be NULL", -1);
         return;
     }
-    UNUSED_PARAMETER(argc);
     int errcode = 0;
     // booster - init
     BoosterHandle booster = NULL;